+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.8.0-dev
+  **NEW FEATURES**
+  - [#174](https://github.com/Datatamer/unify-client-python/issues/174) Get and create taxonomy categories
 
 ## 0.7.0
   **BREAKING CHANGES**